+++ core-ng/src/main/java/core/framework/impl/web/HTTPServerErrorHandler.java
@@ -8,6 +8,8 @@ import core.framework.impl.web.request.RequestImpl;
 import core.framework.impl.web.response.ResponseHandler;
 import core.framework.impl.web.response.ResponseImpl;
 import core.framework.impl.web.service.ErrorResponse;
+import core.framework.impl.web.service.WebServiceClient;
+import core.framework.impl.web.site.AJAXErrorResponse;
 import core.framework.log.ErrorCode;
 import core.framework.log.Severity;
 import core.framework.util.Exceptions;
@@ -15,11 +17,10 @@ import core.framework.web.ErrorHandler;
 import core.framework.web.Response;
 import core.framework.web.service.RemoteServiceException;
 import io.undertow.server.HttpServerExchange;
+import io.undertow.util.HeaderMap;
 import io.undertow.util.Headers;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.util.Optional;
 
 /**
  * @author neo
@@ -41,14 +42,8 @@ public class HTTPServerErrorHandler {
 
         try {
             Response errorResponse = null;
-            if (customErrorHandler != null) {
-                Optional<Response> customErrorResponse = customErrorHandler.handle(request, e);
-                if (customErrorResponse.isPresent()) errorResponse = customErrorResponse.get();
-            }
-            if (errorResponse == null) {
-                String accept = exchange.getRequestHeaders().getFirst(Headers.ACCEPT);
-                errorResponse = defaultErrorResponse(e, accept, actionLog);
-            }
+            if (customErrorHandler != null) errorResponse = customErrorHandler.handle(request, e).orElse(null);
+            if (errorResponse == null) errorResponse = defaultErrorResponse(e, exchange, actionLog);
             responseHandler.render((ResponseImpl) errorResponse, exchange, actionLog);
         } catch (Throwable error) {
             logger.error(error.getMessage(), e);
@@ -60,13 +55,45 @@ public class HTTPServerErrorHandler {
         }
     }
 
-    private Response defaultErrorResponse(Throwable e, String accept, ActionLog actionLog) {
+    private Response defaultErrorResponse(Throwable e, HttpServerExchange exchange, ActionLog actionLog) {
         HTTPStatus status = httpStatus(e);
 
+        HeaderMap headers = exchange.getRequestHeaders();
+        String accept = headers.getFirst(Headers.ACCEPT);
+
         if (accept != null && accept.contains(ContentType.APPLICATION_JSON.mediaType())) {
-            return Response.bean(errorResponse(e, actionLog.id)).status(status);
+            String userAgent = headers.getFirst(Headers.USER_AGENT);
+            return Response.bean(errorResponse(e, userAgent, actionLog.id)).status(status);
         } else {
             return Response.text(errorHTML(e)).status(status).contentType(ContentType.TEXT_HTML);
+        }
+    }
+
+    Object errorResponse(Throwable e, String userAgent, String actionId) {
+        if (WebServiceClient.USER_AGENT.equals(userAgent)) {
+            var response = new ErrorResponse();
+            response.id = actionId;
+            response.message = e.getMessage();
+            response.stackTrace = Exceptions.stackTrace(e);
+            if (e instanceof ErrorCode) {
+                ErrorCode errorCode = (ErrorCode) e;
+                response.errorCode = errorCode.errorCode();
+                response.severity = errorCode.severity().name();
+            } else {
+                response.errorCode = "INTERNAL_ERROR";
+                response.severity = Severity.ERROR.name();
+            }
+            return response;
+        } else {
+            var response = new AJAXErrorResponse();
+            response.id = actionId;
+            response.message = e.getMessage();
+            if (e instanceof ErrorCode) {
+                response.errorCode = ((ErrorCode) e).errorCode();
+            } else {
+                response.errorCode = "INTERNAL_ERROR";
+            }
+            return response;
         }
     }
 
@@ -86,21 +113,5 @@ public class HTTPServerErrorHandler {
         exchange.setStatusCode(HTTPStatus.INTERNAL_SERVER_ERROR.code);
         actionLog.context("responseCode", exchange.getStatusCode());
         exchange.getResponseSender().send(errorHTML(e));
-    }
-
-    ErrorResponse errorResponse(Throwable e, String actionId) {
-        ErrorResponse response = new ErrorResponse();
-        response.id = actionId;
-        response.message = e.getMessage();
-        response.stackTrace = Exceptions.stackTrace(e);
-        if (e instanceof ErrorCode) {
-            ErrorCode errorCode = (ErrorCode) e;
-            response.errorCode = errorCode.errorCode();
-            response.severity = errorCode.severity().name();
-        } else {
-            response.errorCode = "INTERNAL_ERROR";
-            response.severity = Severity.ERROR.name();
-        }
-        return response;
     }
 }