+++ core-ng/src/main/java/core/framework/impl/web/site/AJAXErrorResponse.java
@@ -1,0 +1,17 @@
+package core.framework.impl.web.site;
+
+import core.framework.api.json.Property;
+
+/**
+ * @author neo
+ */
+public class AJAXErrorResponse {
+    @Property(name = "id")
+    public String id;
+
+    @Property(name = "errorCode")   // for ajax, camel case is more js friendly
+    public String errorCode;
+
+    @Property(name = "message")
+    public String message;
+}