+++ core-ng/src/test/java/core/framework/impl/web/HTTPServerErrorHandlerTest.java
@@ -2,6 +2,8 @@ package core.framework.impl.web;
 
 import core.framework.api.http.HTTPStatus;
 import core.framework.impl.web.service.ErrorResponse;
+import core.framework.impl.web.service.WebServiceClient;
+import core.framework.impl.web.site.AJAXErrorResponse;
 import core.framework.log.Severity;
 import core.framework.web.exception.NotFoundException;
 import core.framework.web.service.RemoteServiceException;
@@ -33,19 +35,36 @@ class HTTPServerErrorHandlerTest {
 
     @Test
     void errorResponseWithErrorCodeException() {
-        ErrorResponse response = handler.errorResponse(new NotFoundException("test-message", "TEST_ERROR_CODE"), null);
+        var expected = new ErrorResponse();
+        expected.id = "actionId";
+        expected.message = "test message";
+        expected.errorCode = "TEST_ERROR_CODE";
+        expected.severity = "WARN";
 
-        assertThat(response.message).isEqualTo("test-message");
-        assertThat(response.errorCode).isEqualTo("TEST_ERROR_CODE");
-        assertThat(response.severity).isEqualTo("WARN");
+        var response = handler.errorResponse(new NotFoundException("test message", "TEST_ERROR_CODE"), WebServiceClient.USER_AGENT, "actionId");
+        assertThat(response).isInstanceOf(ErrorResponse.class).isEqualToIgnoringGivenFields(expected, "stackTrace");
     }
 
     @Test
     void errorResponse() {
-        ErrorResponse response = handler.errorResponse(new Error("test-message"), null);
+        var expected = new ErrorResponse();
+        expected.id = "actionId";
+        expected.message = "test message";
+        expected.errorCode = "INTERNAL_ERROR";
+        expected.severity = "ERROR";
 
-        assertThat(response.message).isEqualTo("test-message");
-        assertThat(response.errorCode).isEqualTo("INTERNAL_ERROR");
-        assertThat(response.severity).isEqualTo("ERROR");
+        var response = handler.errorResponse(new Error("test message"), WebServiceClient.USER_AGENT, "actionId");
+        assertThat(response).isInstanceOf(ErrorResponse.class).isEqualToIgnoringGivenFields(expected, "stackTrace");
+    }
+
+    @Test
+    void ajaxErrorResponse() {
+        var expected = new AJAXErrorResponse();
+        expected.id = "actionId";
+        expected.message = "test message";
+        expected.errorCode = "INTERNAL_ERROR";
+
+        var response = handler.errorResponse(new Error("test message"), "Mozilla/5.0", "actionId");
+        assertThat(response).isInstanceOf(AJAXErrorResponse.class).isEqualToIgnoringGivenFields(expected, "stackTrace");
     }
 }