+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Change log
 ### 6.3.7 (6/21/2018)
 * api: make RemoteServiceException exposes https status
+* site: provide default ajax/json error response handling to cover common cases 
 
 ### 6.3.6 (6/14/2018 - 6/21/2018)
 * api: assert same service interface must not have duplicate method name